this neuron will find a linear function
In this problem i shall implement a single neuron that learns to regress (approximate) a function. This function will be linear, so i can use a single linear neuron.
My algorithm reads a number of 2-valued training examples: each such example consists of one input ("x") and a desired output value ("y"). The exact number of training examples is unknown, but you can safely assume you will read <= 1000.
At some point the program will find a training example '0,0\n' , (note that this input might be a possible training point, but we define that (0,0) is invalid!). This indicates that the training data is completely read, and the program should start training the neuron.

After training, the program continues to read 1-valued evaluation data: for each such example the program should compute and print the neurons output followed by return ('\n').


Note: the data (0,0) is not included in my training set!